+++ tests/test_compare.py
@@ -1,5 +1,6 @@
 import pytest
 
+import semver
 from semver import VersionInfo, compare
 
 
@@ -291,7 +292,7 @@ def test_should_not_allow_to_compare_version_with_int():
     with pytest.raises(TypeError):
         1 > v1
     with pytest.raises(TypeError):
-        v1.compare(1)
+        semver.compare(1)
 
 
 def test_should_compare_prerelease_and_build_with_numbers():
