+++ tox.ini
@@ -4,6 +4,7 @@ envlist =
     py{36,37,38,39,310}
     docs
     mypy
+isolated_build = True
 
 
 [testenv]
@@ -35,14 +36,14 @@ commands = flake8 {posargs:}
 description = Check code style
 basepython = python3
 deps = mypy
-commands = mypy {posargs:--ignore-missing-imports .}
+commands = mypy {posargs:--ignore-missing-imports --check-untyped-defs src}
 
 
 [testenv:docstrings]
 description = Check for PEP257 compatible docstrings
 basepython = python3
 deps = docformatter
-commands = docformatter --check {posargs:--pre-summary-newline semver.py}
+commands = docformatter --check {posargs:--pre-summary-newline -r src}
 
 
 [testenv:checks]
@@ -94,3 +95,4 @@ deps =
     git+https://github.com/twisted/towncrier.git
 commands =
     towncrier {posargs:build --draft}
+
